+++ design-toolkit/components/src/components/link/styles.tsx
@@ -16,9 +16,9 @@ export const LinkStyles = tv({
   base: [
     'inline-flex cursor-pointer items-center gap-[0.5ch] px-[0.5ch] align-bottom',
     'enabled:fg-accent-primary-bold',
-    'enabled:hover:fg-accent-primary-hover enabled:hover:bg-interactive-bold-hover',
-    'enabled:focus-visible:fg-accent-primary-hover enabled:focus-visible:bg-interactive-bold-hover',
-    'enabled:pressed:fg-accent-primary-pressed enabled:pressed:bg-interactive-bold-pressed',
+    'enabled:hover:fg-accent-primary-hover enabled:hover:bg-accent-primary-muted',
+    'enabled:focus-visible:fg-accent-primary-hover enabled:focus-visible:bg-accent-primary-muted',
+    'enabled:pressed:fg-accent-primary-pressed enabled:pressed:bg-accent-primary-muted',
     'disabled:fg-disabled disabled:cursor-not-allowed',
   ],
 });